backend/requirements.txt
# Web Framework & Server
fastapi
uvicorn[standard]
python-multipart
pydantic[email]
pydantic-settings

# Database
sqlalchemy
psycopg2-binary
pgvector
alembic

# Security & Auth
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
fastapi-mail

# AI & Plagiarism Logic
sentence-transformers
torch
networkx
numpy
nltk

# PDF Processing
PyMuPDF

# Testing & Utilities
requests
pytest
httpx